refactor test cases into separate files and add missing tests
