+++ py/torch_migraphx/fx/tracer/acc_tracer/acc_ops.py
@@ -1,7 +1,7 @@
 import operator
 import warnings
 
-import torch  # isort:skip
+import torch
 from typing import cast, Iterable, List, Sequence
 
 import torch.nn as nn
